First pass at containerizing unit tests
Run `docker-compose build` to try it out.

We should probabaly create a separate GitHub Actions workflow to run the
tests so we can have that as a separate check that runs prior to the
deploy workflow. Another benefit of this approach is that we can be much
more liberal with the triggers for the tests workflow (i.e. any push).