Overhaul feedback to user (#45)
* Link results to FOLIO Orders & Inventory apps #19

* PO -> /orders/view/{poUUID}
* POL -> /orders/lines/view/{polUUID}
* Title -> /inventory/view/{instanceUUID}

* Drop special font; rename response -> results

* PO Line # from GET response, not internal map #32

The follow up GET to `/orders/composite-orders/{id}` includes both the
`poLineNumber` and `instanceId` per PO Line in the response [1].

This ensures we report back the correct instance associated with each PO
Line in the results block rendered to the user. All that remains is to
now correctly report the instance Title as well.

[1]: https://s3.amazonaws.com/foliodocs/api/mod-orders/p/order.html#orders_composite_orders__id__get

* Display requester in results when applicable #37

* New baseFolioUrl environment variable

* Render internal & receiving notes in results

* Use instance title from composite order GET #32

Ensures POL and Instance reported in feedback to user are completely in
sync. Hat tip to @fereira for pointing out the `titleOrPackage` key in
the response.

He plans on refactoring this in a followup PR, creating an interator for
the `updatedPurchaseOrderJson.compositePoLines` JSONArray in lieu of
iterating through the records in the submitted MARC file a second time.

* Render the Instance HRID (001) in the results

As per Natalya's request.

* Link up URLs & emails embedded within notes

Courtesy of Autolinker.js [1] via Handlebars Custom Helper [2] and HTML
Escaping [3].

[1]: https://github.com/gregjacobs/Autolinker.js
[2]: https://handlebarsjs.com/guide/#custom-helpers
[3]: https://handlebarsjs.com/guide/#html-escaping

* Increase width of primary content area

* Clear the results between submissions